Moved workerclient & workerserver from api/ to pkg/
