+++ pkg/workerapi/workerserver/model/response/artifact.go
